fix: pmm2

+++ arch/i386/pmm/pmm.h
@@ -6,7 +6,7 @@
 
 #define PAGE_SIZE 4096
 
-void pmm_init(uint32_t mb_addr);
+void pmm_init(uintptr_t mb_addr);
 void* pmm_alloc_block(void);
 void pmm_free_block(void* ptr);
